Nodes hide according to checkboxes.  SVG Background is retained.
The edges do not hide yet, and the nodes do not come back after
unchecking the boxes, but the structure for both of those features
are there.  The svg also does not lose it's background, which was a
product of blanking the class attribute of the graph element. Because
a class change is likely to be necessary, the code is still there,
only commented out.