Edges disappear with node on filter.
When a checkbox for a filter is clicked, the edges disappear as well as the nodes.  This is done by temporarily setting all edges (selected as "line") to have the "active" class and then
assigning certain ones as "inactive" if the nodes they are attatched to are also inactive.  This seems expensive, but not enough that it would be a problem for any graph that is
reasonably sized.  If extremely large graphs are to be supported, then refactoring should be done.

An alternative method would be to assign the edges the node_type as well, and thus selection and class assignment would be handled by the same code that handles the hiding of the nodes.
This would lessen the amount of code as well as scale better as there is less written to the dom and less logic being done. The chosen method was selected over this one because it
removes the reliance of edges on the node itself, and thus does not rely on specific structures in the dataset, in theory making it more configurable and easily extendable in the future.