All edges filter regardless of target
This commit fixes #25

The sample data used for development had certain edges which had the root node as the target instead of the source.  While this is technically a problem with the data in this instance, a similar thing would happen when you have more than one root node, which will occur often.  Instead of matching just the target-id to the node to check if it is inactive, it now will check both.  Because this can easily be done with the (amazingly convenient) W3C contains selector (in the format of [attribute*=thingItContains] I was also able to remove the target-id attribute from the edges entirely.