Separate zoom and drag mouse actions
The previous attempts did not take into account that the scale and translate were both being driven by the same event handler.  With this commit, I seperated them out logically, but not
into their own handlers.  I think that multiple handlers (one for dragging and one for zooming) are a bit overkill since they affect the same attribute on the same element.  It might be
best to change the name of the function from zoom() to translate for clarity.  I've left it as is for now, but see it as a likely change in the future but it will be left until I've
asked Huston why they were in the same function in the first place, just in case.

Adjust zoom amount

Purely Aesthetic