Update documentation instructions
Because we are moving the documentation to the gh-pages branch, instructions on compiling the docs is not needed in the master README.  Furthermore, it is now set up to use grunt, and
thus `mkdocs serve` is no longer accurate.  In it's place is a placeholder link that will refer to where gh-pages is hosted, as well as a quick reference to where the documentation is
housed incase anyone wants to help flesh out the docs.