Allow svg to render before testing
This feels super hacky, but at the current draw rate treating
alchemy.begin() as asynchronous and setting a timer to allow
it a full second to run is the only way to make sure what we
are testing is even loaded.

Obviously we need to optimize.  The 1 second on the timer is
arbitrary and more than enough time.  It will do it currently
for the small sample dataset with a 200ms timeout, but until
we optimize, it's better safe than sorry.

Research on others who have had this problem, who are far and
few between, show that they too have used this solution
regardless of how uneasy the hackiness makes them.