deselect all on background svg click
This is slow.  It works for now but there needs to be a better way to do
this, or at least some way to optimize it a bit.  Not noticeable on
smaller graphs, but a thousand edges seems to be the threshold of
noticeability in lag.  Haven't tested threshold of usability.