Do not recreate quadtree function everytime
D3 quadtree works by generating a function that can be called on each
tick.  We were creating a new quadtree every tick.  Instead I've moved
it up to the layout constructor so a new one is made only when we make a
significant enough change to the layout to make a new one.