currentRelationshipTypes{} scoped within alchemy
Fixes #312

...well, for now anyway.
The fact that we have it stored this way is sort of a code smell, but
a few things that will all be extracted out into plugins in 0.4 rely on
it.  For this reason I merely scoped it within the alchemy instance, and
plan on removing it entirely when the plugins are extracted and
refactored.  Things that will be effected are stats and filters.