Fix event reset of initial scale
The problem was introduced with the actual merging of the user
configuration with the defaults at runtime instead of merely comparing
them.  Though this method (introduced in 0.2.2) allows for a greater
degree of flexibility and greatly reduces complexity by calling only one
configuration file, there were still two places (setting of initial
scale and initial translate) where we were comparing them to the default
configuration instead of just applying what the end result of the
configuration was.  This has now been rectified, and both work as
described.

This commit fixes #439